Updated the Project Graphs list behavior as follows:
When an item is selected and the user clicks in an empty row (below the last list item), the selection is dropped.
When an item is selected and the user clicks on a GUI control outside of the panel, the focus is lost (any selection dims).
When an item is clicked, the mouse event is flagged as "used".

git-svn-id: https://www.detoxstudios.com/projectdepot/Development/Unity@2436 b191eb17-2330-428c-a13f-159e02ee8e09