Added the Hotkeys windows, which shows all of the hotkeys in the game.
Any user input (key and mouse) is reflected in the window by highlighting the shortcuts associated with the input. The main uScript EditorWindow will also pass input events (key and mouse) to the HotkeysWindow for handling by that window.

Added a couple Preference options that are read/set by the uScriptHotkeysWindow which affect window behavior. If the user has set their primary click to the right mouse button in their system, Unity reports the RMB as button 0 and the LMB as button 1. The player can select which mouse button is primary in Preferences so the hotkeys show up correctly in the window.

At this time, the hotkeys window shows all hotkeys. Eventually, it may filter the results so reflect actions that the user can perform based on what they're doing in the uScript editor.

git-svn-id: https://www.detoxstudios.com/projectdepot/Development/Unity@2638 b191eb17-2330-428c-a13f-159e02ee8e09