Created a new Attribute AutoLinkType which allows you to specify what type of variable is connected to a node's parameter when the user chooses "CreateLinkedVariable"
The main feature of this is it removes the hardcoding of specific link types in the editor logic.. and allows users with custom nodes to specify the link types for each parameter.  The currently used example is auto linking a scalar when the parameter is an array.

Example:  AddInt is AddInt(int[] a, int[] b).  But CreateLinkedVariable should auto link a single Int node, not an Int List
This removes issue: http://www.uscript.net/forum/viewtopic.php?f=9&t=2324&p=19259#p19259

  

git-svn-id: https://www.detoxstudios.com/projectdepot/Development/Unity@2641 b191eb17-2330-428c-a13f-159e02ee8e09