Completely re-wrote the UpdateNotification class.
It used to utilize native C# calls to make WWW requests and process the responses. Occasionally, the request would timeout (after a couple minutes) and other times it would hang Unity entirely. Since these calls were made during the StartupCheck to see if a new uScript version was available online, this had the potential to significantly delay uScript launch time (or require the user to manually kill the Unity process). Part of the problem is that I believe the WebRequest calls are meant to be run in their own thread, which isn't easily done (AFAIK) in Unity.

The new system uses Unity's WWW class, which is actually intended for runtime web requests, because they usually rely on Coroutines, which do not work in the Edior. As an alternative, I was able to use the EditorWindow Update loop to keep checking until the WWW request had completed. This allows for an asynchronous web request, which doesn't block Unity at all. If the request is taking too long, the user an cancel it via a Cancel button.

Unfortunately, I couldn't utilize the EditorWindow Update loop for the silent StartupCheck, because I didn't want the window to appear unless an update was actually available. So that web request is still synchronous (blocking), but I added a timeout of 5 seconds, so it will not block for more than 5 seconds. Usually the request completes in about 0.3 seconds for me.

Lastly, the update notifications now use an EditorWindow instead of OS-like DisplayDialog popups. I'll send screenshots of what these look like tomorrow, if you're interested.

+ These changes were tested on Windows Unity 3.5.7 and 4.6.1

git-svn-id: https://www.detoxstudios.com/projectdepot/Development/Unity@2689 b191eb17-2330-428c-a13f-159e02ee8e09