I forgot that you cannot use the null coalescing operator (??) to test for null GameObjects, because when they're destroyed, they aren't really null objects.
Unity overloads the == operator to return true for destroyed GameObjects. This "fake null" allows tests like "(GO == null)" to be TRUE when the GO is destroyed. Unfortunately, the ?? operator uses a pure C# null check, which identifies a destroyed GO as NOT null, although all of its members are null, and you can't actually use it any longer in Unity.

Oddly, the ?? use worked in Unity 3 and 4, but not 5. Regardless, we shouldn't use it to test whether a GameObject is null (whether uninitialized or destroyed).

http://blogs.unity3d.com/2014/05/16/custom-operator-should-we-keep-it/

git-svn-id: https://www.detoxstudios.com/projectdepot/Development/Unity@2731 b191eb17-2330-428c-a13f-159e02ee8e09