Fixed a few issues under Unity 3 and Unity 5.
Regarding the 'Content Icon Viewer', it uses a Unity utility method called IconContent(string), which returns an internal texture by its name. When a texture is not found by the name, it spits out a console error message, which cannot be avoided easily. The errors are HARMLESS and SHOULD BE IGNORED. The errors are a result of different versions of Unity including different texture names.

At this time, it is explicitly searching for some texture that exist only in Unity 4.6. Therefore, you'll see an "Unable to load 'xxx'" message for each missing texture under Unity 3 and 5.

git-svn-id: https://www.detoxstudios.com/projectdepot/Development/Unity@2735 b191eb17-2330-428c-a13f-159e02ee8e09