Added the ability to refresh or update the reflected types.
At the bottom of the Toolbox panel (just above the Favorites panel, if that is visible), the user can see the current number of reflected types in the scene. The user may press the "Refresh" button to update the reflected types or enable the "Auto" button to have the list automatically updated whenever the scene Hierarchy changes. The "Auto" setting is maintained across sessions.