The uScript EditorWindow has (2x) anti-alias enabled when the canvas has a zoom scale other than 1.0. The text is a bit easier to read when the graph is zoomed out and the text decreases in size.
Anti-aliasing is disabled when it returns to the normal/default zoom of 1.0, which I assume will offer better performance.

I didn't see any noticeable visual or performance difference between 2x and 8x anti-alias, although we can set it to 2x, 4x, 8x, or off. If users report sluggishness, we can add a preference to enable/disable the feature, or allow them to specify the level they want.