Updated the way that the SceneObjectPathField and AutoCompletePopup cache data. This addresses a text entry slow down that would occur with scenes that contain tens of thousands of objects.
Text entry should now be very quick, although there may be a 1-2 second pause after clicking some variable nodes that utilize the SceneObjectPathField. This delay may be addressed in a future update, but it's far better than the previous text entry slowdown which might have taken up to 5 seconds for each of the first few entered letters into the field.