Implement window "re-attachment" when child windows stay open after the uScript main window has closed (and subsequently re-opens).
uScript.cs/uScriptGUIPanelWindow.cs: Turn panel visibility fields into full properties so that we can do things when they change.  Do not close all sub-windows when uScript closes.  Reattach any already open uScript windows when launching.  Add WeakInstance property which just returns the instance static var, but doesn't re-create it if it's null.
uScriptGUIPanel.cs: Turn InUScriptPanel into full property so it can be overridden as needed.  Add DrawOrphanNotification() which is to be used for a sub-window when the main uScript window is not open.
uScriptGUIPanelContent/uScriptGUIPanelToolbox: Override InUScriptPanel to automatically sync this property between these panels.  In Draw(), use WeakInstance to determine if the main uScript window is open or not and change the window contents accordingly.  Don't set mouse region if InUScriptPanel is false.  When switching to contents/toolbox from current, if we're not in the main uScript panel, find the existing window and switch its Panel and title to the other of the pair.
uScriptGUIPanelProperty.cs/uScriptGUIPanelReference.cs/PanelScript.cs/PanelScriptCurrent.cs/PanelScriptList.cs: In Draw(), use WeakInstance to determine if the main uScript window is open or not and change the window contents accordingly.  Don't set mouse region if InUScriptPanel is false.